requirements.txt
Flask==3.0.0
flask-cors==4.0.0
OPi.GPIO==0.5.0
